requirements.txt
django==5.0.3
djangorestframework==3.15.0
djoser==2.2.2
mysqlclient==2.2.4
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0